Title: 범위 내에 있는 점의 수 2, Time: 232ms, Memory: 16MB, Status: Wrong Answer - CodeTree

+++ 240913/범위 내에 있는 점의 수 2/the-number-of-points-within-the-range-2.cpp
@@ -2,7 +2,7 @@
 #include <algorithm>
 using namespace std;
 
-int a[2000003];
+long a[2000003];
 long sum[2000003];
 int main() 
 {
